feat(headerbox): implement new api for sliver header box
