+++ lib/src/header_box/header_box.dart
@@ -1,8 +1,8 @@
-export 'sbb_cascade_column.dart';
 export 'sbb_header_box.dart';
 export 'sbb_header_box_flap.dart';
-export 'sbb_sliver_floating_header_box.dart';
 export 'sbb_sliver_header_box.dart';
-export 'theme/sbb_header_box_style.dart';
+export 'sbb_sliver_header_box_spacer.dart';
+export 'sliver/sbb_cascade_column.dart';
 export 'theme/sbb_header_box_flap_style.dart';
+export 'theme/sbb_header_box_style.dart';
 export 'theme/sbb_header_box_theme_data.dart';
