--- lib/src/header_box/render_sliver_pin_header.dart
@@ -1,32 +1,0 @@
-import 'dart:math';
-
-import 'package:flutter/rendering.dart';
-
-class RenderSliverPinHeader extends RenderSliverSingleBoxAdapter {
-  @override
-  void performLayout() {
-    child!.layout(constraints.asBoxConstraints(), parentUsesSize: true);
-    double childExtent;
-    switch (constraints.axis) {
-      case .horizontal:
-        childExtent = child!.size.width;
-        break;
-      case .vertical:
-        childExtent = child!.size.height;
-        break;
-    }
-    final paintedChildExtent = min(childExtent, constraints.remainingPaintExtent - constraints.overlap);
-    geometry = SliverGeometry(
-      paintExtent: paintedChildExtent,
-      maxPaintExtent: childExtent,
-      maxScrollObstructionExtent: childExtent,
-      paintOrigin: constraints.overlap,
-      scrollExtent: childExtent,
-      layoutExtent: max(0.0, paintedChildExtent - constraints.scrollOffset),
-      hasVisualOverflow: paintedChildExtent < childExtent,
-    );
-  }
-
-  @override
-  double childMainAxisPosition(_) => 0;
-}
