+++ lib/src/header_box/sliver/sbb_cascade_column.listener.dart
@@ -9,7 +9,7 @@ typedef ContractionListenerBuilder =
 
 /// A widget that listens to expansion state changes.
 ///
-/// It allows you to react to changes in the nearest [SBBCascadeColumn] or [SBBSliverFloatingHeaderBox] and thus must
+/// It allows you to react to changes in the nearest [SBBCascadeColumn] or [SBBSliverHeaderBox] and thus must
 /// must be a descendant of either of them.
 ///
 /// You can provide a [child] which will be passed into [builder] which can be beneficial for performance.
